requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Werkzeug==2.3.6
Flask-Migrate==4.0.4
Flask-Login==0.6.2
bcrypt==4.0.1